Changed exercise id from long to int

+++ src/com/cwru/controller/PerformWorkout.java
@@ -153,7 +153,7 @@ public class PerformWorkout extends FragmentActivity {
 		String exerciseSequence = mDbHelper.getExerciseSequence(workoutName);
 		StringTokenizer st = new StringTokenizer(exerciseSequence,",");
 		while (st.hasMoreTokens()) {
-			Long exerciseId =  Long.parseLong(st.nextToken());
+			int exerciseId = Integer.parseInt(st.nextToken());
 			Exercise exercise = mDbHelper.getExerciseFromId(exerciseId);
 			exercisesForWorkout.add(exercise);
 		}
